pyproject.toml
[tool.poetry]
name = "ascii-art"
version = "0.1.0"
description = "Make ascii from an image"
authors = ["trumully <truman.mulholland@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
hypothesis = {extras = ["cli"], version = "^6.100.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"